Add OrderedMap as map keeping insertion order
We add OrderedMap which  wraps around a Go map keeping the order
with which elements have been added. Keys must be strings, but values
can be anything (interface{}).
Unlike map, index assigment is not possible. Use the `Set` method to
set a key with a particular value.
Use the Keys method to retrieves keys, Values to get the values. To
get both, which probably what you want, use the KeysValues method.