Add ServeReMux multiplexer using regular expressions
ServeMux is an HTTP request multiplexer which matches the path
of the URL of each incoming request against a list of
registered regular expressions.

Every registered can also operationally associated with allowed
HTTP methods. When no allowed method is provided, it will
only allow GET.

When no regular expression matched, 404 is returned. If a pattern
matched, but it turns out the method was not allowed, the HTTP status
405 (method not allowed) is returned.