Add captures to ServeReMux multiplexer
A pattern can now also capture values using `<>` angle brackets.

For example, `/blog/<blogUID>` will capture the value where `<blogUID>`
and it will be made available in the request's context under the key
`xhttp.CapturesContextKey` as an instance of `xhttp.Capture`.

It is also possible to add a type, for example `<int:blogID>`, so
that the methods `AsInt` and `AsInt64` of `xhttp.Capture` can be used
to get the converted value.
It is not possible to use the same named capture twice.