Add package xdoc for generating documentation from Go source
We add a new package xdoc which contains APIs and a command line
interface (CLI) for generating documentation, mimicking pkg.go.dev.

The intention is to have something that can generate documentation
for private/commercial projects. Functionality like a server serving
the docs (like `go doc`) is not available, however user can implement
this herself.

Though not complete, it is already functional. Some parts are still
missing like types created from other types (such as `type Foo string`).

The index generate is optimized for GitHub, but the header links might
work on other platforms and IDEs.

The CLI can be used as follows:

    go run github.com/geertjanvdk/xkit/xdoc/cmd/xgodoc \
        --package-dir path/to/your/pkg -output pkg_doc.md