Add xos.EnvFromStruct reading EnvVar using struct tags
We add functionality which allows defining tags in a struct so that
when passed to `xos.EnvFromStruct`, the fields are populated with
values from the environment.

For example, the following struct:

    type Config struct {
        Address   string `envVar:"ADDRESS" default:"127.0.0.1"`
        DebugMode bool `envVar:"DEBUG"`
    }

    cfg := &Config{}
    EnvFromStruct(cfg)

This will result in:
* when `ADDRESS` is available as environment variable, its value
  will be assigned to `cfg.Address`. If not, 'default' is used.
* When `DEBUG` is set in the environment, the `cfg.DebugMode` is set
  to `true` (value can be 't' or 'true' or '1', or just set)

Environment variables which are empty string, are not considered except
when the field is of `bool`-type.